// src/types/Car.ts
export interface Car {
  id: number | bigint;
  Make: string;
  Model: string;
  Generation: string;
  Year_from: number;
  Year_to: number;
  Series: string;
  Trim: string;
  Body_type: string;
  load_height_mm: number;
  number_of_seats: number;
  length_mm: number;
  width_mm: number;
  height_mm: number;
  wheelbase_mm: number;
  front_track_mm: number;
  rear_track_mm: number;
  curb_weight_kg: number;
  wheel_size_r14: string;
  ground_clearance_mm: number;
  trailer_load_with_brakes_kg: number;
  payload_kg: number;
  back_track_width_mm: number;
  front_track_width_mm: number;
  clearance_mm: number;
  full_weight_kg: number;
  front_rear_axle_load_kg: number;
  max_trunk_capacity_l: number;
  cargo_compartment_length_width_height_mm: string;
  cargo_volume_m3: number;
  minimum_trunk_capacity_l: number;
  maximum_torque_n_m: number;
  injection_type: string;
  overhead_camshaft: string;
  cylinder_layout: string;
  number_of_cylinders: number;
  compression_ratio: number;
  engine_type: string;
  valves_per_cylinder: number;
  boost_type: string;
  cylinder_bore_mm: number;
  stroke_cycle_mm: number;
  engine_placement: string;
  cylinder_bore_and_stroke_cycle_mm: string;
  turnover_of_maximum_torque_rpm: number;
  max_power_kw: number;
  presence_of_intercooler: boolean;
  capacity_cm3: number;
  engine_hp: number;
  engine_hp_rpm: number;
  drive_wheels: string;
  bore_stroke_ratio: number;
  number_of_gears: number;
  turning_circle_m: number;
  transmission: string;
  mixed_fuel_consumption_per_100_km_l: number;
  range_km: number;
  emission_standards: string;
  fuel_tank_capacity_l: number;
  acceleration_0_100_km_h_s: number;
  max_speed_km_per_h: number;
  city_fuel_per_100km_l: number;
  CO2_emissions_g_km: number;
  fuel_grade: string;
  highway_fuel_per_100km_l: number;
  back_suspension: string;
  rear_brakes: string;
  front_brakes: string;
  front_suspension: string;
  steering_type: string;
  car_class: string;
  country_of_origin: string;
  number_of_doors: number;
  safety_assessment: string;
  rating_name: string;
  battery_capacity_KW_per_h: number;
  electric_range_km: number;
  charging_time_h: number;
}